moved IDefaultOptions to interface.ts
